+++ Dust.Restful.Core/Repositories/Interfaces/IRepository.cs
@@ -9,11 +9,11 @@ namespace Dust.Restful.Core.Repositories.Interfaces
     public interface IRepository<T> where T : DataModel, new()
     {
         public List<T> GetAll(int row);
-        public T Get(int id);
-        public bool Exist(int id);
-        public void Delete(int id);
+        public T Get(long id);
+        public bool Exist(long id);
+        public void Delete(long id);
         public bool Insert(T data);
-        public bool Insert(T data, out int id);
+        public bool Insert(T data, out long id);
         public T GetLast();
         public bool Edit(T data);
     }
